backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.2
pydantic-settings==2.1.0
databricks-sql-connector==3.0.2
databricks-sdk==0.18.0
pyyaml==6.0.1
jinja2==3.1.2
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.25.2
structlog==23.2.0
rich==13.7.0
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
mypy==1.7.1
requests==2.31.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# SQL API dependencies
sqlparse==0.4.4
asyncpg==0.29.0
pyparsing==3.1.1